module: cfi: ensure __cfi_check alignment
For modules with no executable code, LLVM generates a __cfi_check
stub, but won't align it to page size as expected. This change ensures
the function is at the beginning of the .text section and correctly
aligned for the CFI shadow.

Also discard the .eh_frame section, which LLD may emit with CFI_CLANG.

Link: https://bugs.llvm.org/show_bug.cgi?id=46293
Signed-off-by: Sami Tolvanen <samitolvanen@google.com>
Signed-off-by: Diab Neiroukh <lazerl0rd@thezest.dev>